Struct bluerobotics_ping::message::ProtocolMessage

pub struct ProtocolMessage {
    pub payload_length: u16,
    pub message_id: u16,
    pub src_device_id: u8,
    pub dst_device_id: u8,
    pub payload: Vec<u8>,
    pub checksum: u16,
}

Fields
payload_length: u16message_id: u16src_device_id: u8dst_device_id: u8payload: Vec<u8>checksum: u16
Implementations

impl ProtocolMessage

pub fn new() -> Self

Message Format

Each message consists of a header, optional payload, and checksum. The binary format is specified as follows:

Byte	Type	Name	Description

0	u8	start1	Start frame identifier, ASCII ‘B’

1	u8	start2	Start frame identifier, ASCII ‘R’

2-3	u16	payload_length	Number of bytes in payload.

4-5	u16	message_id	The message id.

6	u8	src_device_id	The device ID of the device sending the message.

7	u8	dst_device_id	The device ID of the intended recipient of the message.

8-n	u8[]	payload	The message payload.

(n+1)-(n+2)	u16	checksum	The message checksum. The checksum is calculated as the sum of all the non-checksum bytes in the message.

pub fn set_message(&mut self, message: &impl PingMessage)

pub fn set_src_device_id(&mut self, src_device_id: u8)

pub fn dst_device_id(&self) -> u8

pub fn set_dst_device_id(&mut self, dst_device_id: u8)

pub fn payload(&self) -> &[u8]

pub fn checksum(&self) -> u16

pub fn update_checksum(&mut self)

pub fn calculate_crc(&self) -> u16

pub fn has_valid_crc(&self) -> bool

pub fn length(&self) -> usize

pub fn write(&self, writer: &mut dyn Write) -> Result<usize>

pub fn serialized(&self) -> Vec<u8>