bluerobotics_ping

Module omniscan450

Structs

Device

JsonWrapperStruct

No documentation provided.

OsMonoProfileStruct

A profile produced from a single acoustic measurement. The data returned is an array of response strength, in dB, at even intervals across the scan region. The scan region is defined as the region between <scan_start> and <scan_start + scan_length> millimeters away from the transducer. Note that the amplitude of the results are in dB. This is just to keep high dynamic range in the u16 sized data elements. Normally these are converted back to linear power or signal levels.

OsPingParamsStruct

No documentation provided.

PingProtocolHead

SetSpeedOfSoundStruct

Set the speed of sound.

Enums

Messages