bluerobotics_ping

Module ping1d

Structs

ContinuousStartStruct

Command to initiate continuous data stream of profile messages.

ContinuousStopStruct

Command to stop the continuous data stream of profile messages.

Device

DeviceIdStruct

The device ID.

DistanceSimpleStruct

The distance to target with confidence estimate.

DistanceStruct

The distance to target with confidence estimate. Relevant device parameters during the measurement are also provided.

FirmwareVersionStruct

Device information

GainSettingStruct

The current gain setting.

GeneralInfoStruct

General information.

GotoBootloaderStruct

Send the device into the bootloader. This is useful for firmware updates.

ModeAutoStruct

The current operating mode of the device. Manual mode allows for manual selection of the gain and scan range.

PcbTemperatureStruct

Temperature of the on-board thermistor.

PingEnableStruct

Acoustic output enabled state.

PingIntervalStruct

The interval between acoustic measurements.

PingProtocolHead

ProcessorTemperatureStruct

Temperature of the device cpu.

ProfileStruct

A profile produced from a single acoustic measurement. The data returned is an array of response strength at even intervals across the scan region. The scan region is defined as the region between <scan_start> and <scan_start + scan_length> millimeters away from the transducer. A distance measurement to the target is also provided.

RangeStruct

The scan range for acoustic measurements. Measurements returned by the device will lie in the range (scan_start, scan_start + scan_length).

SetDeviceIdStruct

Set the device ID.

SetGainSettingStruct

Set the current gain setting.

SetModeAutoStruct

Set automatic or manual mode. Manual mode allows for manual selection of the gain and scan range.

SetPingEnableStruct

Enable or disable acoustic measurements.

SetPingIntervalStruct

The interval between acoustic measurements.

SetRangeStruct

Set the scan range for acoustic measurements.

SetSpeedOfSoundStruct

Set the speed of sound used for distance calculations.

SpeedOfSoundStruct

The speed of sound used for distance calculations.

TransmitDurationStruct

The duration of the acoustic activation/transmission.

Voltage5Struct

The 5V rail voltage.

Enums

Messages