Module bluerobotics_ping::bluebps

Structs

CellTimeoutStruct

Get the undervoltage timeout

CellVoltageMinStruct

Get the minimum allowed cell voltage

CurrentMaxStruct

get the maximum allowed battery current

CurrentTimeoutStruct

Get the over-current timeout

Device

EraseFlashStruct

Erase flash, including parameter configuration and event counters. The mcu has a limited number of write/erase cycles (1k)!

EventsStruct

A record of events causing a power lock-out. These numbers are non-volatile and reset only with the erase_flash control message.

PingProtocolHead

RebootStruct

reboot the system

ResetDefaultsStruct

Reset parameter configuration to default values.

SetCellVoltageMinimumStruct

Set the minimum allowed cell voltage

SetCellVoltageTimeoutStruct

Set the under-voltage timeout

SetCurrentMaxStruct

Set the maximum allowed battery current

SetCurrentTimeoutStruct

Set the over-current timeout

SetLpfSampleFrequencyStruct

the frequency to take adc samples and run the filter.

SetLpfSettingStruct

Low pass filter setting. This value represents x in the equation value = value * x + sample * (1-x). 0.0 = no filtering, 0.99 = heavy filtering.

SetStreamRateStruct

Set the frequency to automatically output state messages.

SetTemperatureMaxStruct

Set the maximum allowed battery temperature

SetTemperatureTimeoutStruct

Set the over-temperature timeout

StateStruct

Get the current state of the device

TemperatureMaxStruct

Get the maximum allowed battery temperature

TemperatureTimeoutStruct

Get the over-temperature timeout

Enums

Messages